requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.8 # Apache-2.0
pyyaml
dockerfile-parse
attrs
futures
tqdm
docker
